Added transformExtensions option for sourcemaps
ManifestPlugin currently doesn't handle sourcemap assets properly.
"main.js.map" and "main.css.map" will result in a single output key,
"main.map". This is caused by the expectation of only a single file
extension. This is probably not the best way to solve this, but it
works with minimal code changes. The imageExtensions RegExp was also
changed to handle this. Additional extensions can easily be added
for other modifiers.