added module-asset compilation step
Non-images and other non-whitelisted file extensions
(fonts, videos) aren't being included in the manifest. This removes
the imageExtension whitelist, but is a non-BC break because it
preserves the behavior. It's also more future-proof because it
doesn't rely on the whitelist. In addition, this fixes the mapping
of files with additional dots in their names.