Fix mini-css-extract-plugin 0.4.3 issue
Second attempt. Based on https://github.com/webdeveric/webpack-assets-manifest/pull/40

Resolves https://github.com/danethurber/webpack-manifest-plugin/issues/167

mini-css-extract-plugin reports additional, incorrect information for files that are refenced in CSS. The first time we see the file the `module.userRequest` is correct, and we add to `moduleAssets` correctly. However mini-css-extract-plugin then also reports the same `file` but with `module.userRequest` set to the CSS file that references it, which caused us to overwrite the good value in `moduleAssets`.

See the change in mini-css-extract-plugin that caused this https://github.com/webpack-contrib/mini-css-extract-plugin/pull/177