feat: useLegacyEmit option (#246)
* Use same hook for both webpack 4 and 5

* Add useLegacyEmit option

This option is used as compat option for older plugins.

* Add Unit test for useLegacyEmit option

This commit adds a new unit test in the options section that tests the new useLegacyEmit option.
The test runs DependencyExtractionWebpackPlugin that is locked on version 3.1.0. This version emits the main.asset.json on the emit hook.
If useLegacyEmit is set to false the test fails on the webpack v5 tests.